Renamed from `pyEDAA.Configure` to `pyEDAA.ToolSetup`.
