+++ .github/workflows/Pipeline.yml
@@ -11,7 +11,7 @@ jobs:
   Params:
     uses: pyTooling/Actions/.github/workflows/Parameters.yml@r0
     with:
-      name: pyEDAA.Configure
+      name: pyEDAA.ToolSetup
 
   UnitTesting:
     uses: pyTooling/Actions/.github/workflows/UnitTesting.yml@r0
@@ -40,7 +40,7 @@ jobs:
       requirements: '-r tests/requirements.txt'
       commands: |
         cd pyEDAA
-        mypy --html-report ../htmlmypy -p Configure
+        mypy --html-report ../htmlmypy -p ToolSetup
       report: 'htmlmypy'
       artifact: ${{ fromJson(needs.Params.outputs.params).artifacts.typing }}
