+++ README.md
@@ -1,32 +1,32 @@
 <p align="center">
-  <a title="edaa-org.github.io/pyEDAA.Configuration" href="https://edaa-org.github.io/pyEDAA.Configuration"><img height="80px" src="doc/_static/logo.svg"/></a>
+  <a title="edaa-org.github.io/pyEDAA.ToolSetup" href="https://edaa-org.github.io/pyEDAA.ToolSetup"><img height="80px" src="doc/_static/logo.svg"/></a>
 </p>
 
-[![Sourcecode on GitHub](https://img.shields.io/badge/pyEDAA-Configuration-ffca28.svg?longCache=true&style=flat-square&logo=GitHub&labelColor=ff8f00)](https://GitHub.com/edaa-org/pyEDAA.Configuration)
-[![Documentation](https://img.shields.io/website?longCache=true&style=flat-square&label=edaa-org.github.io%2FpyEDAA.Configuration&logo=GitHub&logoColor=fff&up_color=blueviolet&up_message=Read%20now%20%E2%9E%9A&url=https%3A%2F%2Fedaa-org.github.io%2FpyEDAA.Configuration%2Findex.html)](https://edaa-org.github.io/pyEDAA.Configuration/)
+[![Sourcecode on GitHub](https://img.shields.io/badge/pyEDAA-ToolSetup-ffca28.svg?longCache=true&style=flat-square&logo=GitHub&labelColor=ff8f00)](https://GitHub.com/edaa-org/pyEDAA.ToolSetup)
+[![Documentation](https://img.shields.io/website?longCache=true&style=flat-square&label=edaa-org.github.io%2FpyEDAA.ToolSetup&logo=GitHub&logoColor=fff&up_color=blueviolet&up_message=Read%20now%20%E2%9E%9A&url=https%3A%2F%2Fedaa-org.github.io%2FpyEDAA.ToolSetup%2Findex.html)](https://edaa-org.github.io/pyEDAA.ToolSetup/)
 [![Gitter](https://img.shields.io/badge/chat-on%20gitter-4db797.svg?longCache=true&style=flat-square&logo=gitter&logoColor=e8ecef)](https://gitter.im/hdl/community)  
-[![GitHub Workflow - Build and Test Status](https://img.shields.io/github/workflow/status/edaa-org/pyEDAA.Configuration/Pipeline/main?longCache=true&style=flat-square&label=Build%20and%20Test&logo=GitHub%20Actions&logoColor=FFFFFF)](https://GitHub.com/edaa-org/pyEDAA.Configuration/actions/workflows/Pipeline.yml)
+[![GitHub Workflow - Build and Test Status](https://img.shields.io/github/workflow/status/edaa-org/pyEDAA.ToolSetup/Pipeline/main?longCache=true&style=flat-square&label=Build%20and%20Test&logo=GitHub%20Actions&logoColor=FFFFFF)](https://GitHub.com/edaa-org/pyEDAA.ToolSetup/actions/workflows/Pipeline.yml)
 
 <!--
-[![Sourcecode License](https://img.shields.io/pypi/l/pyEDAA.Configuration?longCache=true&style=flat-square&logo=Apache&label=code)](LICENSE.md)
+[![Sourcecode License](https://img.shields.io/pypi/l/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=Apache&label=code)](LICENSE.md)
 [![Documentation License](https://img.shields.io/badge/doc-CC--BY%204.0-green?longCache=true&style=flat-square&logo=CreativeCommons&logoColor=fff)](LICENSE.md)
 
-[![PyPI](https://img.shields.io/pypi/v/pyEDAA.Configuration?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)](https://pypi.org/project/pyEDAA.Configuration/)
-![PyPI - Status](https://img.shields.io/pypi/status/pyEDAA.Configuration?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)
-![PyPI - Python Version](https://img.shields.io/pypi/pyversions/pyEDAA.Configuration?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)
+[![PyPI](https://img.shields.io/pypi/v/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)](https://pypi.org/project/pyEDAA.ToolSetup/)
+![PyPI - Status](https://img.shields.io/pypi/status/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)
+![PyPI - Python Version](https://img.shields.io/pypi/pyversions/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=PyPI&logoColor=FBE072)
 
-[![Libraries.io status for latest release](https://img.shields.io/librariesio/release/pypi/pyEDAA.Configuration?longCache=true&style=flat-square&logo=Libraries.io&logoColor=fff)](https://libraries.io/github/edaa-org/pyEDAA.Configuration)
-[![Codacy - Quality](https://img.shields.io/codacy/grade/39d312bf98244961975559f141c3e000?longCache=true&style=flat-square&logo=Codacy)](https://app.codacy.com/gh/edaa-org/pyEDAA.Configuration)
-[![Codacy - Coverage](https://img.shields.io/codacy/coverage/39d312bf98244961975559f141c3e000?longCache=true&style=flat-square&logo=Codacy)](https://app.codacy.com/gh/edaa-org/pyEDAA.Configuration)
-[![Codecov - Branch Coverage](https://img.shields.io/codecov/c/github/edaa-org/pyEDAA.Configuration?longCache=true&style=flat-square&logo=Codecov)](https://codecov.io/gh/edaa-org/pyEDAA.Configuration)
+[![Libraries.io status for latest release](https://img.shields.io/librariesio/release/pypi/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=Libraries.io&logoColor=fff)](https://libraries.io/github/edaa-org/pyEDAA.ToolSetup)
+[![Codacy - Quality](https://img.shields.io/codacy/grade/39d312bf98244961975559f141c3e000?longCache=true&style=flat-square&logo=Codacy)](https://app.codacy.com/gh/edaa-org/pyEDAA.ToolSetup)
+[![Codacy - Coverage](https://img.shields.io/codacy/coverage/39d312bf98244961975559f141c3e000?longCache=true&style=flat-square&logo=Codacy)](https://app.codacy.com/gh/edaa-org/pyEDAA.ToolSetup)
+[![Codecov - Branch Coverage](https://img.shields.io/codecov/c/github/edaa-org/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=Codecov)](https://codecov.io/gh/edaa-org/pyEDAA.ToolSetup)
 
-[![Dependent repos (via libraries.io)](https://img.shields.io/librariesio/dependent-repos/pypi/pyEDAA.Configuration?longCache=true&style=flat-square&logo=GitHub)](https://GitHub.com/edaa-org/pyEDAA.Configuration/network/dependents)
-[![Requires.io](https://img.shields.io/requires/github/edaa-org/pyEDAA.Configuration?longCache=true&style=flat-square)](https://requires.io/github/EDAA-ORG/pyEDAA.Configuration/requirements/?branch=main)
-[![Libraries.io SourceRank](https://img.shields.io/librariesio/sourcerank/pypi/pyEDAA.Configuration?longCache=true&style=flat-square)](https://libraries.io/github/edaa-org/pyEDAA.Configuration/sourcerank)
+[![Dependent repos (via libraries.io)](https://img.shields.io/librariesio/dependent-repos/pypi/pyEDAA.ToolSetup?longCache=true&style=flat-square&logo=GitHub)](https://GitHub.com/edaa-org/pyEDAA.ToolSetup/network/dependents)
+[![Requires.io](https://img.shields.io/requires/github/edaa-org/pyEDAA.ToolSetup?longCache=true&style=flat-square)](https://requires.io/github/EDAA-ORG/pyEDAA.ToolSetup/requirements/?branch=main)
+[![Libraries.io SourceRank](https://img.shields.io/librariesio/sourcerank/pypi/pyEDAA.ToolSetup?longCache=true&style=flat-square)](https://libraries.io/github/edaa-org/pyEDAA.ToolSetup/sourcerank)
 -->
 
 <p align="center">
-  <a title="edaa-org.github.io/pyEDAA.Configuration" href="https://edaa-org.github.io/pyEDAA.Configuration"><img height="275px" src="doc/_static/work-in-progress.png"/></a>
+  <a title="edaa-org.github.io/pyEDAA.ToolSetup" href="https://edaa-org.github.io/pyEDAA.ToolSetup"><img height="275px" src="doc/_static/work-in-progress.png"/></a>
 </p>
 
 
@@ -86,7 +86,7 @@ class ToolInformation:
 
 ```python
 from pathlib import Path
-from pyEDAA.Configure import Installations
+from pyEDAA.ToolSetup import Installations
 
 yamlFile = Path("configuration.yml")
 
@@ -106,7 +106,7 @@ print(activeHDLVersion.BinaryDirectory)
 * [Patrick Lehmann](https://github.com/Paebbels) (Maintainer)
 * [Martin Zabel](https://github.com/mzabeltud)
 * [Unai Martinez-Corral](https://github.com/umarcor)
-* [and more...](https://github.com/edaa-org/pyEDAA.Configuration/graphs/contributors)
+* [and more...](https://github.com/edaa-org/pyEDAA.ToolSetup/graphs/contributors)
 
 
 # License
