+++ doc/index.rst
@@ -3,7 +3,7 @@
 .. image:: _static/logo.svg
    :height: 90 px
    :align: center
-   :target: https://GitHub.com/edaa-org/pyEDAA.Configuration
+   :target: https://GitHub.com/edaa-org/pyEDAA.ToolSetup
 
 .. raw:: html
 
@@ -15,27 +15,27 @@
 
 .. only:: html
 
-   |  |SHIELD:svg:Configuration-github| |SHIELD:svg:Configuration-ghp-doc| |SHIELD:svg:Configuration-gitter|
-   |  |SHIELD:svg:Configuration-gha-test|
+   |  |SHIELD:svg:ToolSetup-github| |SHIELD:svg:ToolSetup-ghp-doc| |SHIELD:svg:ToolSetup-gitter|
+   |  |SHIELD:svg:ToolSetup-gha-test|
 
-.. Disabled shields: |SHIELD:svg:Configuration-src-license| |SHIELD:svg:Configuration-doc-license| |SHIELD:svg:Configuration-pypi-tag| |SHIELD:svg:Configuration-pypi-status| |SHIELD:svg:Configuration-pypi-python| |SHIELD:svg:Configuration-lib-status| |SHIELD:svg:Configuration-codacy-quality| |SHIELD:svg:Configuration-codacy-coverage| |SHIELD:svg:Configuration-codecov-coverage| |SHIELD:svg:Configuration-lib-dep| |SHIELD:svg:Configuration-req-status| |SHIELD:svg:Configuration-lib-rank|
+.. Disabled shields: |SHIELD:svg:ToolSetup-src-license| |SHIELD:svg:ToolSetup-doc-license| |SHIELD:svg:ToolSetup-pypi-tag| |SHIELD:svg:ToolSetup-pypi-status| |SHIELD:svg:ToolSetup-pypi-python| |SHIELD:svg:ToolSetup-lib-status| |SHIELD:svg:ToolSetup-codacy-quality| |SHIELD:svg:ToolSetup-codacy-coverage| |SHIELD:svg:ToolSetup-codecov-coverage| |SHIELD:svg:ToolSetup-lib-dep| |SHIELD:svg:ToolSetup-req-status| |SHIELD:svg:ToolSetup-lib-rank|
 
 .. only:: latex
 
-   |SHIELD:png:Configuration-github| |SHIELD:png:Configuration-ghp-doc| |SHIELD:png:Configuration-gitter|
-   |SHIELD:png:Configuration-gha-test|
+   |SHIELD:png:ToolSetup-github| |SHIELD:png:ToolSetup-ghp-doc| |SHIELD:png:ToolSetup-gitter|
+   |SHIELD:png:ToolSetup-gha-test|
 
-.. Disabled shields: |SHIELD:png:Configuration-src-license| |SHIELD:png:Configuration-doc-license| |SHIELD:png:Configuration-pypi-tag| |SHIELD:png:Configuration-pypi-status| |SHIELD:png:Configuration-pypi-python| |SHIELD:png:Configuration-lib-status| |SHIELD:png:Configuration-codacy-quality| |SHIELD:png:Configuration-codacy-coverage| |SHIELD:png:Configuration-codecov-coverage| |SHIELD:png:Configuration-lib-dep| |SHIELD:png:Configuration-req-status| |SHIELD:png:Configuration-lib-rank|
+.. Disabled shields: |SHIELD:png:ToolSetup-src-license| |SHIELD:png:ToolSetup-doc-license| |SHIELD:png:ToolSetup-pypi-tag| |SHIELD:png:ToolSetup-pypi-status| |SHIELD:png:ToolSetup-pypi-python| |SHIELD:png:ToolSetup-lib-status| |SHIELD:png:ToolSetup-codacy-quality| |SHIELD:png:ToolSetup-codacy-coverage| |SHIELD:png:ToolSetup-codecov-coverage| |SHIELD:png:ToolSetup-lib-dep| |SHIELD:png:ToolSetup-req-status| |SHIELD:png:ToolSetup-lib-rank|
 
-The pyEDAA.Configuration Documentation
-######################################
+The pyEDAA.ToolSetup Documentation
+##################################
 
 *TBD*
 
 .. image:: _static/work-in-progress.png
    :height: 275 px
    :align: center
-   :target: https://GitHub.com/edaa-org/pyEDAA.Configuration
+   :target: https://GitHub.com/edaa-org/pyEDAA.ToolSetup
 
 .. raw:: html
 
@@ -49,7 +49,7 @@ Contributors
 * `Patrick Lehmann <https://GitHub.com/Paebbels>`__ (Maintainer)
 * `Martin Zabel <https://GitHub.com/https://github.com/mzabeltud>`__
 * `Unai Martinez-Corral <https://GitHub.com/umarcor>`__
-* `and more... <https://GitHub.com/edaa-org/pyEDAA.Configuration/graphs/contributors>`__
+* `and more... <https://GitHub.com/edaa-org/pyEDAA.ToolSetup/graphs/contributors>`__
 
 .. toctree::
    :hidden:
